requirements.txt
pytesseract
Flask
torch
opencv-python
Pillow
NumPy